Altered feature: Todo model, repository, service, controller, and dtos.
Changes in the Todo model: included Serializable, changed done to isDone, and included an account attribute.
Change in the Todo repository: added the query findTodosByAccountId.
Changes in the Todo service: new business logic.
Changes in the Todo controller: included swagger documentation and responses.
Created the TodoRequestDto with the fields: name, description, and priority.
Created the TodoResponseDto with the fields: id, name, description, isDone, and priority.